cleanup config.go until it exists
